Move non-layout Storage related components to `components/interfaces`
