+++ apps/studio/components/interfaces/Storage/AnalyticBucketDetails/CopyEnvButton.tsx
