+++ apps/studio/components/interfaces/Storage/AnalyticBucketDetails/NamespaceRow.tsx
