+++ apps/studio/components/interfaces/Storage/AnalyticBucketDetails/constants.ts
