+++ apps/studio/components/interfaces/Storage/AnalyticBucketDetails/index.tsx
