+++ apps/studio/components/interfaces/Storage/DeleteBucketModal.tsx
