+++ apps/studio/components/interfaces/Storage/EditBucketModal.tsx
@@ -5,11 +5,11 @@ import { useEffect, useState } from 'react'
 import { toast } from 'sonner'
 import { Button, Collapsible, Form, Input, Listbox, Modal, Toggle, cn } from 'ui'
 
-import { StorageSizeUnits } from 'components/to-be-cleaned/Storage/StorageSettings/StorageSettings.constants'
+import { StorageSizeUnits } from 'components/interfaces/Storage/StorageSettings/StorageSettings.constants'
 import {
   convertFromBytes,
   convertToBytes,
-} from 'components/to-be-cleaned/Storage/StorageSettings/StorageSettings.utils'
+} from 'components/interfaces/Storage/StorageSettings/StorageSettings.utils'
 import { InlineLink } from 'components/ui/InlineLink'
 import { useProjectStorageConfigQuery } from 'data/config/project-storage-config-query'
 import { useBucketUpdateMutation } from 'data/storage/bucket-update-mutation'
