+++ apps/studio/components/interfaces/Storage/StorageBucketsError.tsx
