+++ apps/studio/components/interfaces/Storage/StorageExplorer/ColumnContextMenu.tsx
