+++ apps/studio/components/interfaces/Storage/StorageExplorer/ConfirmDeleteModal.tsx
