+++ apps/studio/components/interfaces/Storage/StorageExplorer/FileExplorerRow.tsx
