+++ apps/studio/components/interfaces/Storage/StorageExplorer/FileExplorerRowEditing.tsx
