+++ apps/studio/components/interfaces/Storage/StorageExplorer/FolderContextMenu.tsx
