+++ apps/studio/components/interfaces/Storage/StorageExplorer/PreviewPane.tsx
