+++ apps/studio/components/interfaces/Storage/StorageMenu.tsx
@@ -8,8 +8,8 @@ import { useState } from 'react'
 import { useParams } from 'common'
 import CreateBucketModal from 'components/interfaces/Storage/CreateBucketModal'
 import EditBucketModal from 'components/interfaces/Storage/EditBucketModal'
-import { DeleteBucketModal } from 'components/to-be-cleaned/Storage'
-import { EmptyBucketModal } from 'components/to-be-cleaned/Storage/EmptyBucketModal'
+import { DeleteBucketModal } from 'components/interfaces/Storage'
+import { EmptyBucketModal } from 'components/interfaces/Storage/EmptyBucketModal'
 import { ButtonTooltip } from 'components/ui/ButtonTooltip'
 import ShimmeringLoader from 'components/ui/ShimmeringLoader'
 import { Bucket, useBucketsQuery } from 'data/storage/buckets-query'
