+++ apps/studio/components/interfaces/Storage/StorageSettings/RevokeCredentialModal.tsx
