+++ apps/studio/components/layouts/useLayoutNavCommands.ts
@@ -9,7 +9,7 @@ import { useLogsGotoCommands } from './LogsLayout/Logs.Commands'
 import { useProjectSettingsGotoCommands } from './ProjectSettingsLayout/ProjectSettings.Commands'
 import { useReportsGotoCommands } from './ReportsLayout/Reports.Commands'
 import { useSqlEditorGotoCommands } from './SQLEditorLayout/SqlEditor.Commands'
-import { useStorageGotoCommands } from './StorageLayout/Storage.Commands'
+import { useStorageGotoCommands } from '../interfaces/Storage/Storage.Commands'
 import { useTableEditorGotoCommands } from './TableEditorLayout/TableEditor.Commands'
 
 export function useLayoutNavCommands() {
