+++ apps/studio/pages/project/[ref]/settings/storage.tsx
@@ -6,8 +6,8 @@ import {
   ScaffoldHeader,
   ScaffoldTitle,
 } from 'components/layouts/Scaffold'
-import { StorageSettings } from 'components/to-be-cleaned/Storage'
-import { S3Connection } from 'components/to-be-cleaned/Storage/StorageSettings/S3Connection'
+import { StorageSettings } from 'components/interfaces/Storage'
+import { S3Connection } from 'components/interfaces/Storage/StorageSettings/S3Connection'
 import type { NextPageWithLayout } from 'types'
 
 const PageLayout: NextPageWithLayout = () => {
