+++ apps/studio/pages/project/[ref]/storage/buckets/[bucketId].tsx
@@ -2,11 +2,11 @@ import { useParams } from 'common'
 
 import DefaultLayout from 'components/layouts/DefaultLayout'
 import { useProjectContext } from 'components/layouts/ProjectLayout/ProjectContext'
-import StorageBucketsError from 'components/layouts/StorageLayout/StorageBucketsError'
+import StorageBucketsError from 'components/interfaces/Storage/StorageBucketsError'
 import StorageLayout from 'components/layouts/StorageLayout/StorageLayout'
-import { StorageExplorer } from 'components/to-be-cleaned/Storage'
-import { AnalyticBucketDetails } from 'components/to-be-cleaned/Storage/AnalyticBucketDetails'
-import { useSelectedBucket } from 'components/to-be-cleaned/Storage/StorageExplorer/useSelectedBucket'
+import { StorageExplorer } from 'components/interfaces/Storage'
+import { AnalyticBucketDetails } from 'components/interfaces/Storage/AnalyticBucketDetails'
+import { useSelectedBucket } from 'components/interfaces/Storage/StorageExplorer/useSelectedBucket'
 import { useStorageExplorerStateSnapshot } from 'state/storage-explorer'
 import type { NextPageWithLayout } from 'types'
