+++ apps/studio/pages/project/[ref]/storage/buckets/index.tsx
@@ -1,7 +1,7 @@
 import { useParams } from 'common'
 
 import { useProjectContext } from 'components/layouts/ProjectLayout/ProjectContext'
-import StorageBucketsError from 'components/layouts/StorageLayout/StorageBucketsError'
+import StorageBucketsError from 'components/interfaces/Storage/StorageBucketsError'
 import StorageLayout from 'components/layouts/StorageLayout/StorageLayout'
 import DefaultLayout from 'components/layouts/DefaultLayout'
 import ProductEmptyState from 'components/to-be-cleaned/ProductEmptyState'
