+++ apps/studio/tests/components/Storage/StorageSettings.utils.test.ts
@@ -1,8 +1,8 @@
-import { StorageSizeUnits } from 'components/to-be-cleaned/Storage/StorageSettings/StorageSettings.constants'
+import { StorageSizeUnits } from 'components/interfaces/Storage/StorageSettings/StorageSettings.constants'
 import {
   convertFromBytes,
   convertToBytes,
-} from 'components/to-be-cleaned/Storage/StorageSettings/StorageSettings.utils'
+} from 'components/interfaces/Storage/StorageSettings/StorageSettings.utils'
 import { describe, test, expect } from 'vitest'
 
 describe('StorageSettings.utils: convertFromBytes', () => {
